Move `TypingTarget` related logic out of rule file
